refactor: 검증 메소드 다른 클래스로 분리
